对接口进行调整，已经找到了不能从数据库中删除数据库的问题，原来是使用jpa进行删除数据时，需要在service层添加事务管理，避免删除数据时发生异常

+++ src/main/java/student/course/scsv/service/CourseService.java
@@ -11,9 +11,11 @@ import student.course.scsv.repository.CourseRepository;
 import student.course.scsv.repository.TeacherRepository;
 import student.course.scsv.util.FormatString;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
+@Transactional
 public class CourseService {
 
     @Autowired
@@ -40,8 +42,11 @@ public class CourseService {
      * @return JSON
      */
     public String delCourse(Long cid){
-        courseRepository.deleteByCid(cid);
-        return FormatString.infoToJson("200", "update success");
+        if (courseRepository.exists(cid)){
+            courseRepository.deleteByCid(cid);
+            return FormatString.infoToJson("200", "update success");
+        }
+        return FormatString.infoToJson("400", "this course was not found");
     }
 
     /**
